+++ src/app/components/table/table.component.ts
@@ -1,11 +1,13 @@
 import { AfterViewInit, Component, OnInit, ViewChild } from '@angular/core';
+import { Subscription } from 'rxjs';
 
 import { MatPaginator } from '@angular/material/paginator';
 import { MatTableDataSource } from '@angular/material/table';
 
-import { TABLE_COLUMNS } from './table-columns';
+import { SavingService } from 'src/app/services/saving.service';
 
 import { Saving } from '../../models/saving';
+import { TABLE_COLUMNS } from './table-columns';
 
 @Component({
   selector: 'app-table',
@@ -16,19 +18,19 @@ export class TableComponent implements OnInit, AfterViewInit {
   @ViewChild(MatPaginator) public paginator!: MatPaginator;
 
   public displayedColumns = TABLE_COLUMNS;
-  public dataSource = new MatTableDataSource<Saving>([
-    {
-      id: 1,
-      goal: 'TEST',
-      description: 'Test description',
-      amount: 1000,
-      date: '2024-09-19',
-    },
-  ]);
+  public dataSource = new MatTableDataSource<Saving>([]);
 
-  constructor() {}
+  private subs = new Subscription();
 
-  ngOnInit(): void {}
+  constructor(private savingService: SavingService) {}
+
+  ngOnInit(): void {
+    this.subs.add(
+      this.savingService.savings$.subscribe(
+        (res) => (this.dataSource.data = res)
+      )
+    );
+  }
 
   ngAfterViewInit(): void {
     this.dataSource.paginator = this.paginator;