+++ CHANGELOG.md
@@ -12,3 +12,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Data period selector by specific month or year. Specific date will be add in the future.
 - Table with savings.
 - Service to use API.
+- Service get implementation for period-selector component and table component. Show data on table.